requirements.txt
allure-pytest=2.13.2
pytest=7.4.3
Appium-Python-Client=3.1.1
python-dotenv=1.0.0
selene=2.0.0rc4